Портировать аддон на BBS FS 2.7
Вкладка «Light» в таймлайне регистрируется через новый Addon API
(RegisterTrackCategoriesEvent) вместо дописывания константы в enum
ReplayCategory, которого в 2.7 больше нет. Группы треков лампы переехали
со строк-заголовков частей тела, тоже удалённых, на собственные секции
дорожек 2.7, а тик автоключа стал дробным.

Какой трек чей, BBS теперь спрашивает одним адресом, а свойства лампы
называются как всегда, поэтому по адресу это не решить: раскладка
запоминает адреса треков ламп того каталога, который BBS построила
последним, и правило вкладки читает этот список.

+++ src/client/java/qualet/irlite/client/ui/replays/UIReplaySelectionKeyframeFactory.java
@@ -32,7 +32,7 @@ public final class UIReplaySelectionKeyframeFactory extends UIKeyframeFactory<St
     @Override
     protected String getDisplayValue()
     {
-        Integer tick = this.editor.getGraph().getAutoKeyframeTick();
+        Float tick = this.editor.getGraph().getAutoKeyframeTick();
         UIKeyframeSheet sheet = tick == null ? null : this.editor.getGraph().getSheet(this.keyframe);
 
         if (sheet != null && ReplaySelectionPlayback.startsAfter(sheet.channel, tick))
